feat: ダッシュボードページにタイトルを追加
マイページ配下の各ページに適切なタイトルを設定:
- /dashboard: プロフィール - マイページ - VCharts
- /dashboard/tickets: 応援チケット - マイページ - VCharts
- /dashboard/comments: コメント履歴 - マイページ - VCharts
- /dashboard/subscription: サブスクリプション - マイページ - VCharts
- /dashboard/settings: 通知設定 - マイページ - VCharts

Client Component だった subscription と settings は
Server Component + Client Component の構成に変更
